requirements.txt
django==5.0.0
djangorestframework==3.14.0
Pillow==10.1.0
python-dotenv==1.0.0
django-mptt==0.15.0
django-filter==23.5